Commit 1
Input:

Today we're going to make a blog website. For this conversation, do not
explain what you have made. Only print a single formatted code output in
response. Create a project structure for a boilerplate javascript blog
website.

Output:

Here is a suggested project structure for a boilerplate JavaScript blog
website:

blog-website/
├── css/
│   └── style.css
├── img/
│   └── logo.png
├── js/
│   └── main.js
├── index.html
└── blog/
    └── [blog-post-1].html

This project structure includes a css directory for stylesheets, an img
directory for images, a js directory for JavaScript files, and an
index.html file for the homepage. It also includes a blog directory,
which contains individual HTML files for each blog post.