Commit 8
Input:

Create a templates/ directory in our blog project, then create a template file that can be used to display the attributes from our MDX file.

Output:

Here is a suggested project structure for a blog website that includes a templates directory and a template file for displaying the attributes from an MDX file:

```
blog-website/
├── css/
│   └── style.css
├── img/
│   └── logo.png
├── js/
│   └── main.js
├── templates/
│   └── post.template
├── index.html
└── blog/
    └── [blog-post-1].mdx
```

The templates directory contains a post.template file, which is a template that can be used to display the attributes from an MDX file. This template might include placeholders for the title, author, date, tags, and summary attributes, which can be filled in with the corresponding values from the MDX file.

Here is an example of what the post.template file might look like:

```
<article>
  <h1>{{ title }}</h1>
  <p>by {{ author }} on {{ date }}</p>
  <p>{{ summary }}</p>
  <p>Tags: {{ tags }}</p>
</article>
```

In this example, the template includes placeholders for each of the attributes that can be filled in with the corresponding values from the MDX file. The resulting HTML output would then be a nicely formatted blog post that includes the title, author, date, summary, and tags.