Commit 8
Input:

Create another template in the templates/ directory for the general page layout. The template should have sections for a title bar, a main content area, and a footer bar. The title and footer bars should be fixed to the top of the window, and the content area should be scrollable.

Output:

Here is a suggested project structure for a blog website that includes a templates directory and two template files, one for displaying blog post attributes and one for the general page layout:

```
blog-website/
├── css/
│   └── style.css
├── img/
│   └── logo.png
├── js/
│   └── main.js
├── templates/
│   ├── layout.template
│   └── post.template
├── index.html
└── blog/
    └── [blog-post-1].mdx
```

The templates directory now includes two template files: layout.template and post.template. The layout.template file is a template for the general page layout and includes sections for the title bar, main content area, and footer bar. The post.template file is a template for displaying the attributes from an MDX file.

Here is an example of what the layout.template file might look like:

```
<!DOCTYPE html>
<html>
  <head>
    <title>My Blog</title>
    <link rel="stylesheet" href="css/style.css">
  </head>
  <body>
    <header>
      <h1>My Blog</h1>
    </header>
    <main>
      {{ content }}
    </main>
    <footer>
      <p>Copyright 2021 My Blog</p>
    </footer>
  </body>
</html>
```

In this example, the layout.template file includes a <header> element for the title bar, a <main> element for the main content area, and a <footer> element for the footer bar. The {{ content }} placeholder in the <main> element is where the content from the individual blog post will be inserted.

To make the title and footer bars fixed to the top of the window, you can add the following styles to the css/style.css file:

```
header, footer {
  position: fixed;
  width: 100%;
  top: 0;
  left: 0;
  z-index: 1;
}
```

To make the main content area scrollable, you can add the following styles:

```
main {
  margin-top: 75px;
  margin-bottom: 75px;
  overflow: auto;
}
```

These styles will fix the position of the title and footer bars to the top of the window and make the main content area scrollable, while also leaving enough space at the top and bottom of the content area to avoid it being obscured by the title and footer bars.