Commit 9
Input:

Create a `pages/` subdirectory that will contain the routable pages for our blog. Create an index page in the pages subdirectory that renders a list of blog posts using the layout and post templates.

Output:

Here is a suggested project structure for a blog website that includes a templates directory, a pages directory, and two template files, one for displaying blog post attributes and one for the general page layout:

```
blog-website/
├── css/
│   └── style.css
├── img/
│   └── logo.png
├── js/
│   └── main.js
├── templates/
│   ├── layout.template
│   └── post.template
├── pages/
│   └── index.html
├── index.html
└── blog/
    └── [blog-post-1].mdx
```

The pages directory now contains an index.html file, which is the page that will be displayed when a user navigates to the homepage of the blog. This page will render a list of blog posts using the layout.template and post.template files to provide the general page layout and display the blog post attributes, respectively.

Here is an example of what the pages/index.html file might look like:

```
<!-- Import the layout template -->
<!-- The layout template expects a "content" variable to be defined -->
<% content = `
<!-- Import the post template -->
<!-- The post template expects the following variables to be defined: title,
```