Commit 20
Input:

Update the dist folder with the rendered index page, as well as pages for each blog post

Output:

To update the dist directory with the rendered index page, as well as pages for each blog post, you can create a script that uses a static site generator to parse the content of the MDX files in the blog directory and generate corresponding HTML pages in the dist directory. For example, you could use a tool like next.js or gatsby.js to create the pages, using the templates in the templates directory to define the layout and content of the pages. Once the pages have been generated, you can add them to the dist directory, and they will be available to serve to users of the blog website.