Commit 21
Input:

Show me the contents of the files in dist after being generated by a static site generator

Output:

Here is an example of what the dist directory might look like after being generated by a static site generator, using the content of the MDX files in the blog directory:

```
dist/
├── index.html
├── blog-post-1.html
├── blog-post-2.html
└── blog-post-3.html
```

Each of the HTML files in the dist directory corresponds to a blog post in the blog directory, and contains the fully rendered HTML content of the post, using the templates in the templates directory to define the layout and content of the pages. The index.html file contains the rendered index page, which includes a list of blog posts, as well as links to the individual blog post pages.