Commit 29
Input:

Add a GitHub Action to this project to deploy the static website files in `dist` to GitHub Pages

Output:

To add a GitHub Action to this project that deploys the static website files in dist to GitHub Pages, you can create a .github/workflows directory in your project and add a deploy.yml file to it. The deploy.yml file should contain the following code:

```
name: Deploy to GitHub Pages

on:
  push:
    branches:
      - master

jobs:
  deploy:
    runs-on: ubuntu-latest
    steps:
      - uses: actions/checkout@v2
      - uses: peaceiris/actions-gh-pages@v3
        with:
          github_token: ${{ secrets.GITHUB_TOKEN }}
          publish_dir: ./dist
```

This GitHub Action will be triggered whenever a commit is pushed to the master branch of your repository. When triggered, it will use the peaceiris/actions-gh-pages action to deploy the contents of the dist directory to GitHub Pages.

Once the deploy.yml file has been added to your project, you can push it to your GitHub repository and the action will be triggered automatically. The static website files in dist will be deployed to GitHub Pages, and you will be able to access them at the URL specified in the Settings section of your GitHub repository.