Commit 32
Input:

Update the `.github/workflows/deploy.yml` action to deploy from the `main` branch instead of `master`

Output:

To update the .github/workflows/deploy.yml action to deploy from the main branch instead of master, you can modify the deploy.yml file as follows:

```
name: Deploy to GitHub Pages

on:
  push:
    branches:
      - main

jobs:
  deploy:
    runs-on: ubuntu-latest
    steps:
      - uses: actions/checkout@v2
      - uses: peaceiris/actions-gh-pages@v3
        with:
          github_token: ${{ secrets.GITHUB_TOKEN }}
          publish_dir: ./blog-website/dist
```

This updated deploy.yml file will use the peaceiris/actions-gh-pages action to deploy the contents of the ./blog-website/dist directory to GitHub Pages whenever a commit is pushed to the main branch of your repository. You can then commit this change to your Git repository and push it to GitHub to update the GitHub Action. When the action is triggered, it will use the updated configuration to deploy the static website files from the blog-website/dist directory to GitHub Pages.