Commit 37
Input:

Update the 'Read more' URL links for blog posts to be prefixed with the path `openai-gpt-blog`

Output:

To update the "Read more" URL links for blog posts to be prefixed with the path openai-gpt-blog, you will need to update the code that generates the HTML for the blog post pages. If you are using a static site generator, you can update the configuration or template files that are used to generate the HTML for the blog post pages to include the openai-gpt-blog path in the "Read more" links.

For example, if you are using Next.js to generate the HTML for the blog post pages, you can update the pages/[slug].js file to include the openai-gpt-blog path in the Link component that is used to generate the "Read more" link:

```
import Link from "next/link";

const BlogPostPage = ({ post }) => (
  <Layout>
    <h1>{post.title}</h1>
    <p>{post.summary}</p>
    <p>{post.content}</p>

    {/* Update the Link component to include the openai-gpt-blog path */}
    <Link href="/openai-gpt-blog/[slug]" as={`/openai-gpt-blog/${post.slug}`}>
      <a>Read more</a>
    </Link>
  </Layout>
);
```

Alternatively, if you are using a templating engine to generate the HTML for the blog post pages, you can update the template file to include the openai-gpt-blog path in the "Read more" links. For example, if you are using the EJS templating engine, you can update the blog-post.ejs template file as follows:

```
<h1><%= title %></h1>
<p><%= summary %></p>
<p><%= content %></p>

{/* Update the link to include the openai-gpt-blog path */}
<a href="/openai-gpt-blog/<%= slug %>">Read more</a>
```

Once you have updated the code or templates that generate the HTML for the blog post pages, you can re-render the HTML pages and deploy them to GitHub Pages to update the "Read more" links on your blog website.