Commit 39
Input:

Update the slug URLs to explicitly reference the blog files in the `dist/` directory. For example, the `href` for Post One should be `/openai-gpt-blog/blog-post-1.html`, and the `href` for Post Two should be `/openapi-gpt-blog/blog-post-2.html`

Output:

To update the slug URLs to explicitly reference the blog files in the dist/ directory, you will need to update the code or templates that generate the HTML for the blog post pages to include the blog-post-* file names in the href attributes of the "Read more" links.

For example, if you are using Next.js to generate the HTML for the blog post pages, you can update the pages/[slug].js file to include the blog-post-* file name in the Link component that is used to generate the "Read more" link:

```
import Link from "next/link";

const BlogPostPage = ({ post }) => (
  <Layout>
    <h1>{post.title}</h1>
    <p>{post.summary}</p>
    <p>{post.content}</p>

    {/* Update the Link component to include the blog-post-* file name in the `as` attribute */}
    <Link href="/openai-gpt-blog/blog-post-1.html" as={`/openai-gpt-blog/blog-post-${post.slug}.html`}>
      <a>Read more</a>
    </Link>
  </Layout>
);
```

Alternatively, if you are using a templating engine to generate the HTML for the blog post pages, you can update the template file to include the blog-post-* file name in the href attributes of the "Read more" links. For example, if you are using the EJS templating engine, you can update the blog-post.ejs template file as follows:

```
<h1><%= title %></h1>
<p><%= summary %></p>
<p><%= content %></p>

{/* Update the link to include the blog-post-* file name in the `href` attribute */}
<a href="/openai-gpt-blog/blog-post-1.html">Read more</a>
```

Once you have updated the code or templates that generate the HTML for the blog post pages, you can re-render the HTML pages and deploy them to GitHub Pages to update the "Read more" links on your blog website.